@@ -36,8 +36,8 @@
     "js-cookie": "3.0.5",
     "lodash.debounce": "4.0.8",
     "qs": "6.14.0",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-json-view": "^1.19.1",
     "react-redux": "7.1.0",
     "react-router": "5.0.1",
@@ -51,6 +51,7 @@
     "@babel/plugin-proposal-class-properties": "7.18.6",
     "@babel/preset-env": "7.28.3",
     "@babel/preset-react": "7.27.1",
+    "@types/history": "^4.7.11",
     "babel-cli": "6.26.0",
     "babel-loader": "10.0.0",
     "clean-webpack-plugin": "4.0.0",
