@@ -1,8 +1,10 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
 import 'core-js/stable'
 import 'regenerator-runtime/runtime'
+import { createRoot } from 'react-dom/client'
+import React from 'react'
 
 import App from './views/App'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+const container = document.getElementById('root')
+const root = createRoot(container)
+root.render(<App />)
